move some test files around
